created UI and Junit

+++ src/main/java/com/healthcare/model/Address.java
@@ -1,22 +1,19 @@
 package com.healthcare.model;
 public class Address {
 	
-    private int id;
+    private Long id;
 	
 	
 	 
 	
-    public int getId() {
+    public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	/*
-	 * public int getPatientId() { return patientId; } public void setPatientId(int
-	 * patientId) { this.patientId = patientId; }
-	 */
+
 private String state ;
 private String city;
 private Long pin;
